feature-routes | Refactoring the code to use separate route files for each entity
